# openapi.yaml
...
openapi: 3.0.1
info:
  title: IDS AppStore
  description: IDS AppStore developed by the Fraunhofer FIT
  contact:
    name: IDS AppStore
    url: https://spring.io/projects/spring-boot/appstore
    email: '@email@'
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
  version: 1.0.0
servers:
  - url: https://localhost:8080
    description: Generated server url
...
paths:
...
  /api/notify:
    put:
      tags:
        - Messages
      summary: Notify all subscribers
      description: "Can be used to manually notify all subscribers about a resource\
        \ offer, representation, or artifact update."
      operationId: sendMessage
      parameters:
        - name: elementId
          in: query
          description: The element id.
          required: true
          schema:
            type: string
            format: uri
      responses:
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Not found
          content:
            '*/*':
              schema:
                type: object
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                type: object
        "204":
          description: No Content
          content:
            '*/*':
              schema:
                type: object
...